recompiles again
