+++ examples/add_external/src/main.nr
@@ -3,6 +3,6 @@ pub fn main(
     external: [Field; 2],
     registers: [Field; 2],
     next_pc: Field,
-) -> pub ([Field; 2], Field) {
-    ([external[0] + registers[0], external[1] + registers[1]], next_pc)
+) -> pub (Field, [Field; 2]) {
+    (next_pc, [external[0] + registers[0], external[1] + registers[1]])
 }
