+++ examples/swap_memory/src/main.nr
@@ -1,4 +1,4 @@
 /// Swap the two registers.
-pub fn main(registers: [Field; 2], next_pc: Field) -> pub ([Field; 2], Field) {
-    ([registers[1], registers[0]], next_pc)
+pub fn main(registers: [Field; 2], next_pc: Field) -> pub (Field, [Field; 2]) {
+    (next_pc, [registers[1], registers[0]])
 }
